+++ packages/generator/src/generators/Database.ts
@@ -1,6 +1,7 @@
 /* eslint-disable unicorn/better-regex */
 
 import { stripHexcode } from 'emojibase';
+import { transliterate } from 'transliteration';
 import { log } from '../helpers/log';
 import { toArray } from '../helpers/toArray';
 import { Emoji, EmojiMap, Hexcode, HexcodeMap, ShortcodeDataMap } from '../types';
@@ -26,14 +27,14 @@ export class Database {
 		this.mapEmojis(emojis);
 	}
 
-	static slugify(value: string): string {
+	static slugify(value: string, transform: boolean = false): string {
 		return (
-			value
+			(transform ? transliterate(value) : value)
 				.toLocaleLowerCase()
 				// Apply separators
 				.replace(/(\s|`|\/|\\|･|（|）|／)+/g, '_')
 				// Remove special chars
-				.replace(/([!"&'()[\],.:;<>«»?ʼ’‘“”—–])/g, '')
+				.replace(/([!"&'()[\],.:;<>«»?ʼ’‘“”—–„])/g, '')
 				// Remove multiple underscores
 				.replace(/_{2,}/g, '_')
 				// Remove leading underscores